ebaa-40: Criando controller de auth e alocando rotas para o local correto
