+++ app/Http/Controllers/AuthController.php
@@ -1,0 +1,86 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Repositories\UserRepository;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Tymon\JWTAuth\Exceptions\JWTException;
+
+class AuthController extends Controller
+{
+    public function login(Request $request)
+    {
+        $credentials = $request->only(['email', 'password', 'remember']);
+
+        $remember = (bool) data_get($credentials, 'remember');
+
+        $repository = new UserRepository();
+
+        $user = $repository->getByEmail(data_get($credentials, 'email'));
+
+        if (!$user || !Hash::check(data_get($credentials, 'password'), $user->password)) {
+            abort(401, 'Não autorizado');
+        }
+
+        $token = auth()->login($user, $remember);
+
+        return response()->json([
+            'data' => [
+                'token' => $token,
+                'expires_in' => auth()->factory()->getTTL() * 60 * 8
+            ]
+        ]);
+    }
+
+    public function logout(Request $request)
+    {
+        if (! auth()->user()) {
+            throw new \Exception('Nenhum usuário logado');
+        }
+
+        auth()->logout();
+
+        $request->session()->invalidate();
+
+        $request->session()->regenerateToken();
+
+        return response()->json([
+            'message' => 'Desconectado do sistema'
+        ]);
+    }
+
+    public function me()
+    {
+        if (! auth()->user()) {
+            throw new \Exception('Nenhum usuário logado');
+        }
+
+        return response()->json([
+            'user' => auth()->user()
+        ]);
+    }
+
+    public function refreshToken(Request $request)
+    {
+        try {
+            $token = $request->bearerToken();
+            if (!$token) {
+                return response()->json([
+                    'error' => 'Token invalido'
+                ], 401);
+            }
+
+            $refreshedToken = Auth::guard('api')->refresh($token);
+            return response()->json([
+                'newToken' => $refreshedToken
+            ]);
+        } catch (JWTException $e) {
+            return response()->json([
+                'error' => $e->getMessage()
+            ], 400);
+        }
+    }
+
+}
